python/requirements-dev.txt
aiohttp==3.11.14
aiosignal==1.3.2
anyio==4.9.0
astroid==3.3.9
async-timeout==5.0.1
attrs==25.3.0
azure-core==1.32.0
azure-identity==1.21.0
build==1.2.2.post1
bumpver==2024.1130
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Deprecated==1.2.18
dill==0.3.9
docutils==0.21.2
flit==3.11.0
flit_core==3.11.0
frozenlist==1.5.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.6.1
isort==6.0.1
lazy-object-proxy==1.10.0
lexid==2021.1006
looseversion==1.3.0
mccabe==0.7.0
microsoft-kiota-abstractions==1.9.2
microsoft-kiota-authentication-azure==1.9.2
microsoft-kiota-http==1.9.2
microsoft-kiota-serialization-form==1.9.2
microsoft-kiota-serialization-json==1.9.2
microsoft-kiota-serialization-multipart==1.9.2
microsoft-kiota-serialization-text==1.9.2
-e ./microsoft_agents_m365copilot_core
msal==1.32.0
msal-extensions==1.3.1
multidict==6.2.0
mypy==1.15.0
mypy-extensions==1.0.0
opentelemetry-api==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
packaging==24.2
pathlib2==2.3.7.post1
platformdirs==4.3.6
portalocker==3.1.1
pycparser==2.22
PyJWT==2.10.1
pylint==3.3.4
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
std-uritemplate>=2.0.0
time-machine==2.16.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wrapt==1.17.2
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0